fix: Restore validation by eliminating the dependency loop
