+++ x/deployment/types/order_bid_duration_validation.go
@@ -1,0 +1,1 @@
+package types
